Do not use Exceptions for code control flow for BlockBreakListener
This drastically reduces consumed CPU time when used in conjunction
with mods such as QuarryPlus. This is caused by Exceptions being
rather expensive to initialize. That being said, this mod is still
not as optimized as I'd like to have it be, but at this point all
further expenses are mere rounding errors.